Refactor to move update job application status to separate file due to different routing
