+++ app/routers/joblistingapplication.py
@@ -2,12 +2,10 @@ from typing import Union
 
 from beanie.odm.fields import PydanticObjectId
 from fastapi import APIRouter, HTTPException, status
-from app.constants.application_status import ApplicationStatus
 
 from app.dtos.jobapplication import (GetJobApplicationResDto,
                                      SubmitJobApplicationReqDto,
-                                     SubmitJobApplicationResDto, UpdateJobApplicationStatusReqDto,
-                                     UpdateJobApplicationStatusResDto)
+                                     SubmitJobApplicationResDto)
 from app.models.jobapplication import JobApplication
 from app.models.joblisting import JobListing
 
@@ -41,15 +39,4 @@ async def submit_job_application(job_listing_id: PydanticObjectId, req_dto: Subm
         personalStatement=req_dto.personalStatement,
         jobListing=jobListing  # type: ignore
     )
-    return await jobapplication.save()
-
-
-@router.patch("/{job_application_id}", response_model=UpdateJobApplicationStatusResDto)
-async def update_job_application_status(job_application_id: PydanticObjectId, req_dto: UpdateJobApplicationStatusReqDto):
-    """Updates a job application's status"""
-    jobapplication = await JobApplication.find_one({"_id": job_application_id}, fetch_links=True)
-    if not jobapplication:
-        raise HTTPException(status_code=404, detail="Job application not found")
-
-    jobapplication.applicationStatus = req_dto.applicationStatus
     return await jobapplication.save()
