+++ app/server.py
@@ -9,6 +9,7 @@ from app.models.joblisting import JobListing
 from app.models.user import User
 from app.routers.auth import router as auth_router
 from app.routers.internal import router as internal_router
+from app.routers.jobapplication import router as job_application_router
 from app.routers.joblisting import router as job_listing_router
 from app.routers.joblistingapplication import \
     router as job_listing_application_router
@@ -30,6 +31,7 @@ app.add_middleware(
 
 app.include_router(auth_router)
 app.include_router(internal_router)
+app.include_router(job_application_router)
 app.include_router(job_listing_application_router)
 app.include_router(job_listing_router)
 app.include_router(user_router)
